Fixed ShootableEntity/Environment
Fixed ShootableEntity and Environment scripts so that they work properly. Speed of children scales with time, and environment timeStatus is reduced every 5 seconds. Also added a temporary tester that adjusts the timeStatus of the moving platform with q and e. Moving platform is currently broken though, and needs to be fixed so that the player can stand on it and stay on it properly.