Improved Launcher, Spinner, and Chaser
Launcher now works in the desired way. Can be toggled to only launch when sped up, or launch at all speeds (like if we want a horizontal bumper thing).

Added bumpers to spinner so that anything with a rigidbody gets launched upon hit.

Reworked chaser so that it uses force instead of just setting velocity. So it can now be affected by bumpers or launchers. Also added some wall detection so it will no longer get caught on the side of walls, and improved behavior on slopes.