Added inner hitbox to bullets
Added a small inner hitbox to all bullet types. The main hitbox handles collisions on where the bullet affects what it collides with (player, enemies, phys objects), and the inner hitbox handles other collisions (like walls). This makes it less likely for the bullet's large outer hitbox to collide on the edge of walls.